adds files to common in core
