+++ Blu.core/common/EventLogHelper.cs
@@ -1,12 +1,12 @@
 ﻿using System.Diagnostics;
 
 namespace Blu.core.common
-{
+{ 
     public static class EventLogHelper
     {
         static object eloglock = new object();
         static EventLog elog = new EventLog { Source = Config.ServiceName };
-        
+
         public static void WriteToEventLog(EventLogEntryType type, string message)
         {
             lock (eloglock)
@@ -34,4 +34,3 @@ namespace Blu.core.common
         }
     }
 }
-
