+++ Blu.core/common/StringHelper.cs
