+++ BluService/BluService.csproj
@@ -50,18 +50,6 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="..\BluIpc\Common\Config.cs">
-      <Link>BluIpc\Common\Config.cs</Link>
-    </Compile>
-    <Compile Include="..\BluIpc\Common\EventLogHelper.cs">
-      <Link>BluIpc\Common\EventLogHelper.cs</Link>
-    </Compile>
-    <Compile Include="..\BluIpc\Common\RegistryHelper.cs">
-      <Link>BluIpc\Common\RegistryHelper.cs</Link>
-    </Compile>
-    <Compile Include="..\BluIpc\Common\StringHelper.cs">
-      <Link>BluIpc\Common\StringHelper.cs</Link>
-    </Compile>
     <Compile Include="..\BluIpc\Server\IIpcCallback.cs">
       <Link>BluIpc\Server\IIpcCallback.cs</Link>
     </Compile>
